get embedded album art
